Move share to 3-dot menu
