+++ packages/frontend/src/components/dialogs/ViewProfile/index.tsx
@@ -29,9 +29,6 @@ import type { DialogProps } from '../../../contexts/DialogContext'
 import type { T } from '@deltachat/jsonrpc-client'
 import { RovingTabindexProvider } from '../../../contexts/RovingTabindex'
 import { ChatListItemRowChat } from '../../chat/ChatListItemRow'
-import useCreateDraftMessage from '../../../hooks/chat/useCreateDraftMesssage'
-import { runtime } from '@deltachat-desktop/runtime-interface'
-import SelectChat from '../SelectChat'
 
 const log = getLogger('renderer/dialogs/ViewProfile')
 
@@ -83,7 +80,7 @@ export default function ViewProfile(
 
   const showMenu = !(isDeviceChat || isSelfChat)
 
-  const onClickViewProfileMenu = useViewProfileMenu(contact)
+  const onClickViewProfileMenu = useViewProfileMenu(contact, onClose)
 
   return (
     <Dialog
@@ -242,13 +239,6 @@ export function ViewProfileInner({
 
   const VerificationTag = verifier?.action ? 'button' : 'div'
 
-  const onClickShareContact = () => {
-    onClose()
-    openDialog(ShareProfileDialog, {
-      contact,
-    })
-  }
-
   return (
     <>
       <DialogContent>
@@ -303,9 +293,6 @@ export function ViewProfileInner({
             {tx('menu_unblock_contact')}
           </Button>
         )}
-        <Button styling='secondary' onClick={onClickShareContact}>
-          {tx('menu_share')}
-        </Button>
       </div>
       {statusText !== '' && (
         <>
@@ -363,40 +350,5 @@ export function ViewProfileInner({
         </>
       )}
     </>
-  )
-}
-
-function ShareProfileDialog(props: { contact: T.Contact } & DialogProps) {
-  const { onClose, contact } = props
-
-  const tx = useTranslationFunction()
-  const accountId = selectedAccountId()
-  const createDraftMessage = useCreateDraftMessage()
-
-  const onChatClick = async (chatId: number) => {
-    const vcard = await BackendRemote.rpc.makeVcard(accountId, [contact.id])
-
-    const filePath = await runtime.writeTempFile('contact.vcard', vcard)
-    // treefit: I would like to use setDraftVcard here, but it requires a draft message, which we may now have:
-    // BackendRemote.rpc.setDraftVcard(accountId, msgId, contacts)
-    // and there is no way to create an empty draft message with the current api as far as I know
-    //
-    // why is this better? because we then only would need to ask to replace draft when there is a file
-
-    onClose()
-    await createDraftMessage(accountId, chatId, '', {
-      name: `${contact.displayName}.vcard`,
-      path: filePath,
-    })
-    runtime.removeTempFile(filePath)
-  }
-
-  return (
-    <SelectChat
-      headerTitle={tx('chat_share_with_title')}
-      onChatClick={onChatClick}
-      onClose={onClose}
-      listFlags={C.DC_GCL_FOR_FORWARDING | C.DC_GCL_NO_SPECIALS}
-    />
   )
 }
